feat: add BackToTop button and MiniProjectCard to project details page
