+++ src/lib/browser/post.ts
@@ -1,11 +1,11 @@
+import { listenToBackToTop, showHideBackToTopButton } from './back-to-top';
+
 const shareButtons = document.querySelectorAll(
   '#share'
 ) as NodeListOf<HTMLElement>;
 const shareButtonTexts = document.querySelectorAll(
   '#share-title'
 ) as NodeListOf<HTMLElement>;
-
-const backButton = document.querySelector('.back-to-top') as HTMLLinkElement;
 
 const progressBar = document.querySelector('.progress-bar') as HTMLElement;
 
@@ -27,25 +27,6 @@ function listenToProgressBar() {
   const progress = (window.scrollY / totalHeight) * 100;
 
   progressBar.style.width = `${progress}%`;
-}
-
-/**
- * Go to page top on button click.
- */
-function listenToBackToTop() {
-  backButton.addEventListener('click', () => window.scrollTo(0, 0));
-}
-
-/**
- * Show or hide 'back to top' button after certain pixels.
- */
-function showHideBackToTopButton() {
-  const pageOffset =
-    document.documentElement.scrollTop || document.body.scrollTop;
-
-  pageOffset > 900
-    ? backButton.classList.add('show')
-    : backButton.classList.remove('show');
 }
 
 /**
